+++ rust/chains/hyperlane-cosmos/src/providers/mod.rs
@@ -5,7 +5,7 @@ use hyperlane_core::{
 };
 use tendermint_rpc::{client::CompatMode, HttpClient};
 
-use crate::{ConnectionConf, HyperlaneCosmosError, Signer};
+use crate::{ConnectionConf, CosmosAmount, HyperlaneCosmosError, Signer};
 
 use self::grpc::WasmGrpcProvider;
 
@@ -31,7 +31,8 @@ impl CosmosProvider {
         locator: Option<ContractLocator>,
         signer: Option<Signer>,
     ) -> ChainResult<Self> {
-        let grpc_client = WasmGrpcProvider::new(conf.clone(), locator, signer)?;
+        let gas_price = CosmosAmount::try_from(conf.get_minimum_gas_price().clone())?;
+        let grpc_client = WasmGrpcProvider::new(conf.clone(), gas_price.clone(), locator, signer)?;
         let rpc_client = HttpClient::builder(
             conf.get_rpc_url()
                 .parse()