+++ typescript/infra/config/environments/mainnet3/agent.ts
@@ -43,14 +43,14 @@ const hyperlane: RootAgentConfig = {
     rpcConsensusType: RpcConsensusType.Fallback,
     docker: {
       repo,
-      tag: '36de5bc-20231205-145629',
+      tag: '86b7f98-20231207-153805',
     },
     gasPaymentEnforcement,
   },
   validators: {
     docker: {
       repo,
-      tag: '36de5bc-20231205-145629',
+      tag: '86b7f98-20231207-153805',
     },
     rpcConsensusType: RpcConsensusType.Quorum,
     chains: validatorChainConfig(Contexts.Hyperlane),
@@ -59,7 +59,7 @@ const hyperlane: RootAgentConfig = {
     rpcConsensusType: RpcConsensusType.Fallback,
     docker: {
       repo,
-      tag: '36de5bc-20231205-145629',
+      tag: '86b7f98-20231207-153805',
     },
   },
 };
@@ -72,7 +72,7 @@ const releaseCandidate: RootAgentConfig = {
     rpcConsensusType: RpcConsensusType.Fallback,
     docker: {
       repo,
-      tag: '36de5bc-20231205-145629',
+      tag: '86b7f98-20231207-153805',
     },
     // whitelist: releaseCandidateHelloworldMatchingList,
     gasPaymentEnforcement,
@@ -84,7 +84,7 @@ const releaseCandidate: RootAgentConfig = {
   validators: {
     docker: {
       repo,
-      tag: '36de5bc-20231205-145629',
+      tag: '86b7f98-20231207-153805',
     },
     rpcConsensusType: RpcConsensusType.Quorum,
     chains: validatorChainConfig(Contexts.ReleaseCandidate),
@@ -108,7 +108,7 @@ const neutron: RootAgentConfig = {
     rpcConsensusType: RpcConsensusType.Fallback,
     docker: {
       repo,
-      tag: '36de5bc-20231205-145629',
+      tag: '67585a2-20231220-223937',
     },
     gasPaymentEnforcement: [
       {