+++ typescript/infra/config/environments/testnet4/agent.ts
@@ -50,7 +50,7 @@ const hyperlane: RootAgentConfig = {
     rpcConsensusType: RpcConsensusType.Fallback,
     docker: {
       repo,
-      tag: '36de5bc-20231205-145629',
+      tag: '86b7f98-20231207-153805',
     },
     blacklist: [
       ...releaseCandidateHelloworldMatchingList,
@@ -67,7 +67,7 @@ const hyperlane: RootAgentConfig = {
     rpcConsensusType: RpcConsensusType.Fallback,
     docker: {
       repo,
-      tag: '36de5bc-20231205-145629',
+      tag: '86b7f98-20231207-153805',
     },
     chains: validatorChainConfig(Contexts.Hyperlane),
   },
@@ -75,7 +75,7 @@ const hyperlane: RootAgentConfig = {
     rpcConsensusType: RpcConsensusType.Fallback,
     docker: {
       repo,
-      tag: '36de5bc-20231205-145629',
+      tag: '86b7f98-20231207-153805',
     },
   },
 };
@@ -88,7 +88,7 @@ const releaseCandidate: RootAgentConfig = {
     rpcConsensusType: RpcConsensusType.Fallback,
     docker: {
       repo,
-      tag: '36de5bc-20231205-145629',
+      tag: '86b7f98-20231207-153805',
     },
     whitelist: [...releaseCandidateHelloworldMatchingList],
     gasPaymentEnforcement,
@@ -101,34 +101,13 @@ const releaseCandidate: RootAgentConfig = {
     rpcConsensusType: RpcConsensusType.Fallback,
     docker: {
       repo,
-      tag: '36de5bc-20231205-145629',
+      tag: '86b7f98-20231207-153805',
     },
     chains: validatorChainConfig(Contexts.ReleaseCandidate),
-  },
-};
-
-const neutron: RootAgentConfig = {
-  ...contextBase,
-  context: Contexts.Neutron,
-  rolesWithKeys: [Role.Relayer],
-  contextChainNames: {
-    relayer: [Chains.goerli],
-    validator: [],
-    scraper: [],
-  },
-  relayer: {
-    rpcConsensusType: RpcConsensusType.Fallback,
-    docker: {
-      repo,
-      tag: '36de5bc-20231205-145629',
-    },
-    gasPaymentEnforcement,
-    transactionGasLimit: 750000,
   },
 };
 
 export const agents = {
   [Contexts.Hyperlane]: hyperlane,
   [Contexts.ReleaseCandidate]: releaseCandidate,
-  [Contexts.Neutron]: neutron,
 };