+++ typescript/infra/config/environments/testnet4/chains.ts
@@ -14,7 +14,12 @@ export const ethereumTestnetConfigs: ChainMap<ChainMetadata> = {
       maxPriorityFeePerGas: 40 * 10 ** 9, // 40 gwei
     },
   },
-  bsctestnet: chainMetadata.bsctestnet,
+  bsctestnet: {
+    ...chainMetadata.bsctestnet,
+    transactionOverrides: {
+      gasPrice: 80 * 10 ** 9, // 8 gwei
+    },
+  },
   goerli: chainMetadata.goerli,
   scrollsepolia: chainMetadata.scrollsepolia,
   sepolia: chainMetadata.sepolia,
@@ -27,7 +32,6 @@ export const ethereumTestnetConfigs: ChainMap<ChainMetadata> = {
 // Blessed non-Ethereum chains.
 export const nonEthereumTestnetConfigs: ChainMap<ChainMetadata> = {
   // solanadevnet: chainMetadata.solanadevnet,
-  // neutrontestnet: chainMetadata.neutrontestnet,
 };
 
 export const testnetConfigs: ChainMap<ChainMetadata> = {