+++ typescript/infra/scripts/deploy.ts
@@ -4,6 +4,8 @@ import { prompt } from 'prompts';
 import { HelloWorldDeployer } from '@hyperlane-xyz/helloworld';
 import {
   ChainMap,
+  Chains,
+  HypERC20Deployer,
   HyperlaneCore,
   HyperlaneCoreDeployer,
   HyperlaneDeployer,
@@ -13,7 +15,9 @@ import {
   InterchainAccountDeployer,
   InterchainQueryDeployer,
   LiquidityLayerDeployer,
+  TokenConfig,
 } from '@hyperlane-xyz/sdk';
+import { TokenDecimals, TokenType } from '@hyperlane-xyz/sdk/dist/token/config';
 import { objMap } from '@hyperlane-xyz/utils';
 
 import { Contexts } from '../config/contexts';
@@ -75,6 +79,39 @@ async function main() {
       multiProvider,
     );
     deployer = new HyperlaneCoreDeployer(multiProvider, ismFactory);
+  } else if (module === Modules.WARP) {
+    const owner = deployerAddress;
+    const neutronRouter =
+      '6b04c49fcfd98bc4ea9c05cd5790462a39537c00028333474aebe6ddf20b73a3';
+    const ismFactory = HyperlaneIsmFactory.fromAddressesMap(
+      getAddresses(environment, Modules.PROXY_FACTORY),
+      multiProvider,
+    );
+    const tokenConfig: TokenConfig & TokenDecimals = {
+      type: TokenType.synthetic,
+      name: 'Eclipse Fi',
+      symbol: 'ECLIP',
+      decimals: 6,
+      totalSupply: 0,
+    };
+    const core = HyperlaneCore.fromEnvironment(
+      deployEnvToSdkEnv[environment],
+      multiProvider,
+    );
+    const routerConfig = core.getRouterConfig(owner);
+    const targetChains = [Chains.arbitrum];
+    config = {
+      arbitrum: {
+        ...routerConfig['arbitrum'],
+        ...tokenConfig,
+        interchainSecurityModule: '0x53A5c239d62ff35c98E0EC9612c86517748ffF59',
+        gas: 600_000,
+      },
+      neutron: {
+        foreignDeployment: neutronRouter,
+      },
+    };
+    deployer = new HypERC20Deployer(multiProvider, ismFactory);
   } else if (module === Modules.INTERCHAIN_GAS_PAYMASTER) {
     config = envConfig.igp;
     deployer = new HyperlaneIgpDeployer(multiProvider);