+++ typescript/utils/index.ts
@@ -84,6 +84,7 @@ export {
 export {
   ValueOf,
   arrayToObject,
+  deepCopy,
   deepEquals,
   invertKeysAndValues,
   isObject,