+++ typescript/utils/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @hyperlane-xyz/utils
 
+## 3.5.1
+
+## 3.5.0
+
+## 3.4.0
+
+### Patch Changes
+
+- fd4fc1898: - Upgrade Viem to 1.20.0
+  - Add optional restUrls field to ChainMetadata
+  - Add deepCopy util function
+  - Add support for cosmos factory token addresses
+
 ## 3.3.0
 
 ## 3.2.0